+++ tests/unittests/classification/inputs.py
@@ -270,7 +270,7 @@ _input_multidim_multiclass = Input(
 
 
 # Generate plausible-looking inputs
-def generate_plausible_inputs_multilabel(num_classes=NUM_CLASSES, num_batches=NUM_BATCHES, batch_size=BATCH_SIZE):
+def _generate_plausible_inputs_multilabel(num_classes=NUM_CLASSES, num_batches=NUM_BATCHES, batch_size=BATCH_SIZE):
     correct_targets = torch.randint(high=num_classes, size=(num_batches, batch_size))
     preds = torch.rand(num_batches, batch_size, num_classes)
     targets = torch.zeros_like(preds, dtype=torch.long)
@@ -284,15 +284,15 @@ def generate_plausible_inputs_multilabel(num_classes=NUM_CLASSES, num_batches=NU
     return Input(preds=preds, target=targets)
 
 
-def generate_plausible_inputs_binary(num_batches=NUM_BATCHES, batch_size=BATCH_SIZE):
+def _generate_plausible_inputs_binary(num_batches=NUM_BATCHES, batch_size=BATCH_SIZE):
     targets = torch.randint(high=2, size=(num_batches, batch_size))
     preds = torch.rand(num_batches, batch_size) + torch.rand(num_batches, batch_size) * targets / 3
     return Input(preds=preds / (preds.max() + 0.01), target=targets)
 
 
-_input_multilabel_prob_plausible = generate_plausible_inputs_multilabel()
+_input_multilabel_prob_plausible = _generate_plausible_inputs_multilabel()
 
-_input_binary_prob_plausible = generate_plausible_inputs_binary()
+_input_binary_prob_plausible = _generate_plausible_inputs_binary()
 
 # randomly remove one class from the input
 _temp = torch.randint(high=NUM_CLASSES, size=(NUM_BATCHES, BATCH_SIZE))